backend/requirements.txt
fastapi==0.87.0
spotipy==2.20.0
uvicorn==0.19.0
psycopg2-binary==2.9.3